extracted root entities to specific file
